This commit merges the `test_file_to_file` and `test_file_to_directory` tests of the `test_transfer` into a single test using the `parametrize` decorator of `pytest`. Furthermore, the working directories of tests in `test_remotepath` were changed from the current working directory to the tmp directory

+++ tests/test_transfer.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import asyncio
 import os
 import tempfile
@@ -7,23 +9,21 @@ import pytest
 from streamflow.core import utils
 from streamflow.core.context import StreamFlowContext
 from streamflow.core.data import DataType
-from streamflow.core.deployment import Connector, ExecutionLocation
+from streamflow.core.deployment import ExecutionLocation
 from streamflow.data.remotepath import StreamFlowPath
 from tests.utils.deployment import get_location
 
 
 async def _compare_remote_dirs(
     context: StreamFlowContext,
-    src_connector: Connector,
     src_location: ExecutionLocation,
     src_path: StreamFlowPath,
-    dst_connector: Connector,
     dst_location: ExecutionLocation,
     dst_path: StreamFlowPath,
-):
+) -> None:
     assert await dst_path.exists()
 
-    # the two dirs must have the same elements order
+    # The two directories must have the same order of elements
     src_path, src_dirs, src_files = await src_path.walk(
         follow_symlinks=True
     ).__anext__()
@@ -44,10 +44,8 @@ async def _compare_remote_dirs(
             asyncio.create_task(
                 _compare_remote_dirs(
                     context,
-                    src_connector,
                     src_location,
                     src_path / src_dir,
-                    dst_connector,
                     dst_location,
                     dst_path / dst_dir,
                 )
@@ -57,7 +55,11 @@ async def _compare_remote_dirs(
 
 
 async def _create_tmp_dir(
-    context, connector, location, root=None, lvl=None, n_files=0
+    context: StreamFlowContext,
+    location: ExecutionLocation,
+    root: StreamFlowPath | None = None,
+    lvl: str | None = None,
+    n_files: int = 0,
 ) -> StreamFlowPath:
     dir_lvl = f"-{lvl}" if lvl else ""
     dir_path = StreamFlowPath(
@@ -104,22 +106,17 @@ async def test_directory_to_directory(
     #   |   |   empty
 
     src_location = await get_location(context, communication_pattern[0])
-    src_connector = context.deployment_manager.get_connector(src_location.deployment)
     src_path = None
 
     dst_location = await get_location(context, communication_pattern[1])
-    dst_connector = context.deployment_manager.get_connector(dst_location.deployment)
     dst_path = None
 
     try:
-        # create src structure
-        src_path = await _create_tmp_dir(
-            context, src_connector, src_location, n_files=4
-        )
+        # Create src structure
+        src_path = await _create_tmp_dir(context, src_location, n_files=4)
         for i in range(3):
             inner_dir = await _create_tmp_dir(
                 context,
-                src_connector,
                 src_location,
                 root=src_path,
                 n_files=2 + i if i < 2 else 0,
@@ -128,7 +125,6 @@ async def test_directory_to_directory(
             if i == 0:
                 await _create_tmp_dir(
                     context,
-                    src_connector,
                     src_location,
                     root=inner_dir,
                     n_files=2,
@@ -136,23 +132,19 @@ async def test_directory_to_directory(
                 )
         src_path = await src_path.resolve()
 
-        # dst init
+        # Transfer from `src_path` on `src_location` to `dst_path` directory on `dst_location`
         dst_path = StreamFlowPath(
             tempfile.gettempdir() if dst_location.local else "/tmp",
             utils.random_name(),
             context=context,
             location=dst_location,
         )
-
-        # save src_path into StreamFlow
         context.data_manager.register_path(
             location=src_location,
             path=str(src_path),
             relpath=str(src_path),
             data_type=DataType.PRIMARY,
         )
-
-        # transfer src_path to dst_path
         await context.data_manager.transfer_data(
             src_location=src_location,
             src_path=str(src_path),
@@ -160,28 +152,27 @@ async def test_directory_to_directory(
             dst_path=str(dst_path),
             writable=False,
         )
-
-        # check if dst exists
         await dst_path.exists()
 
-        # check that src and dst have the same sub dirs and files
+        # Check that the source and destination have the same subdirectories and files
         await _compare_remote_dirs(
             context,
-            src_connector,
             src_location,
             src_path,
-            dst_connector,
             dst_location,
             dst_path,
         )
     finally:
-        await src_path.rmtree()
-        await dst_path.rmtree()
+        if src_path is not None:
+            await src_path.rmtree()
+        if dst_path is not None:
+            await dst_path.rmtree()
 
 
 @pytest.mark.asyncio
-async def test_file_to_directory(
-    context: StreamFlowContext, communication_pattern: tuple[str, str]
+@pytest.mark.parametrize("dst_t", ["file", "directory"])
+async def test_file_to_entity(
+    context: StreamFlowContext, dst_t: str, communication_pattern: tuple[str, str]
 ) -> None:
     """Test transferring a file from one location to a directory into another location."""
     src_location = await get_location(context, communication_pattern[0])
@@ -199,10 +190,12 @@ async def test_file_to_directory(
         context=context,
         location=dst_location,
     )
-    await dst_path.mkdir(mode=0o777, exist_ok=True)
+    if dst_t == "directory":
+        await dst_path.mkdir(mode=0o777, exist_ok=True)
     try:
         await src_path.write_text("StreamFlow")
         src_path = await src_path.resolve()
+        assert src_path is not None
         context.data_manager.register_path(
             location=src_location,
             path=str(src_path),
@@ -216,52 +209,9 @@ async def test_file_to_directory(
             dst_path=str(dst_path),
             writable=False,
         )
-        dst_file = dst_path / src_path.name
+        dst_file = dst_path if dst_t == "file" else (dst_path / src_path.name)
         assert await dst_file.exists()
         assert await src_path.checksum() == await dst_file.checksum()
-    finally:
-        await src_path.rmtree()
-        await dst_path.rmtree()
-
-
-@pytest.mark.asyncio
-async def test_file_to_file(
-    context: StreamFlowContext, communication_pattern: tuple[str, str]
-):
-    """Test transferring a file from one location to another."""
-    src_location = await get_location(context, communication_pattern[0])
-    dst_location = await get_location(context, communication_pattern[1])
-
-    src_path = StreamFlowPath(
-        tempfile.gettempdir() if src_location.local else "/tmp",
-        utils.random_name(),
-        context=context,
-        location=src_location,
-    )
-    dst_path = StreamFlowPath(
-        tempfile.gettempdir() if src_location.local else "/tmp",
-        utils.random_name(),
-        context=context,
-        location=dst_location,
-    )
-    try:
-        await src_path.write_text("StreamFlow")
-        src_path = await src_path.resolve()
-        context.data_manager.register_path(
-            location=src_location,
-            path=str(src_path),
-            relpath=str(src_path),
-            data_type=DataType.PRIMARY,
-        )
-        await context.data_manager.transfer_data(
-            src_location=src_location,
-            src_path=str(src_path),
-            dst_locations=[dst_location],
-            dst_path=str(dst_path),
-            writable=False,
-        )
-        assert await dst_path.exists()
-        assert await src_path.checksum() == await dst_path.checksum()
     finally:
         await src_path.rmtree()
         await dst_path.rmtree()
@@ -274,7 +224,9 @@ async def test_multiple_files(
     """Test transferring multiple files simultaneously from one location to another."""
     await asyncio.gather(
         *(
-            asyncio.create_task(test_file_to_file(context, communication_pattern))
+            asyncio.create_task(
+                test_file_to_entity(context, "file", communication_pattern)
+            )
             for _ in range(20)
         )
     )
